+++ LiftLog.Ui/LiftLog.Ui.csproj
@@ -32,6 +32,10 @@
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
             <PrivateAssets>all</PrivateAssets>
         </PackageReference>
+        <PackageReference Include="Microsoft.TypeScript.MSBuild" Version="5.5.3">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
         <PackageReference Include="TaskTupleAwaiter" Version="2.1.0" />
     </ItemGroup>
