+++ LiftLog.Ui/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "strict": true,
+    "noImplicitAny": false,
+    "target": "ES2021",
+    "lib": ["ES2021", "DOM"]
+  }
+}
