+++ LiftLog.Web/wwwroot/index.html
@@ -1,70 +1,73 @@
 <!DOCTYPE html>
 <html>
-<head>
-    <meta charset="utf-8"/>
+  <head>
+    <meta charset="utf-8" />
     <meta
-            name="viewport"
-            content="viewport-fit=cover, width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, user-scalable=no"
+      name="viewport"
+      content="viewport-fit=cover, width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, user-scalable=no"
     />
     <title>LiftLog</title>
-    <base href="/"/>
-    <link href="_content/LiftLog.Ui/css/app.css" rel="stylesheet"/>
-    <link href="_content/LiftLog.Ui/app.min.css" rel="stylesheet"/>
-</head>
+    <base href="/" />
+    <link href="_content/LiftLog.Ui/css/app.css" rel="stylesheet" />
+    <link href="_content/LiftLog.Ui/app.min.css" rel="stylesheet" />
+  </head>
 
-<body>
-<div id="app" class="flex flex-col justify-center h-full text-center select-none"><span>LiftLog</span></div>
+  <body>
+    <div id="app" class="flex flex-col justify-center h-full text-center select-none"><span>LiftLog</span></div>
+    <div id="toast" class="fixed bottom-28 hidden mx-auto flex flex-col items-center right-0 left-0 z-40">
+      <div id="toast-content" class="bg-gray-800 text-white text-center p-2 rounded-lg"></div>
+    </div>
+    <div id="blazor-error-ui">
+      An unhandled error has occurred.
+      <a href="" class="reload">Reload</a>
+      <a class="dismiss">🗙</a>
+    </div>
 
-<div id="blazor-error-ui">
-    An unhandled error has occurred.
-    <a href="" class="reload">Reload</a>
-    <a class="dismiss">🗙</a>
-</div>
-
-<!-- Start Single Page Apps for GitHub Pages -->
-<script type="text/javascript">
-    // Single Page Apps for GitHub Pages
-    // MIT License
-    // https://github.com/rafgraph/spa-github-pages
-    // This script checks to see if a redirect is present in the query string,
-    // converts it back into the correct url and adds it to the
-    // browser's history using window.history.replaceState(...),
-    // which won't cause the browser to attempt to load the new url.
-    // When the single page app is loaded further down in this file,
-    // the correct url will be waiting in the browser's history for
-    // the single page app to route accordingly.
-    (function (l) {
+    <!-- Start Single Page Apps for GitHub Pages -->
+    <script type="text/javascript">
+      // Single Page Apps for GitHub Pages
+      // MIT License
+      // https://github.com/rafgraph/spa-github-pages
+      // This script checks to see if a redirect is present in the query string,
+      // converts it back into the correct url and adds it to the
+      // browser's history using window.history.replaceState(...),
+      // which won't cause the browser to attempt to load the new url.
+      // When the single page app is loaded further down in this file,
+      // the correct url will be waiting in the browser's history for
+      // the single page app to route accordingly.
+      (function (l) {
         if (l.search[1] === "/") {
-            var decoded = l.search
-                .slice(1)
-                .split("&")
-                .map(function (s) {
-                    return s.replace(/~and~/g, "&");
-                })
-                .join("?");
-            window.history.replaceState(null, null, l.pathname.slice(0, -1) + decoded + l.hash);
+          var decoded = l.search
+            .slice(1)
+            .split("&")
+            .map(function (s) {
+              return s.replace(/~and~/g, "&");
+            })
+            .join("?");
+          window.history.replaceState(null, null, l.pathname.slice(0, -1) + decoded + l.hash);
         }
-    })(window.location);
-</script>
+      })(window.location);
+    </script>
 
-<script>
-    // on load try to redirect to the app via liftlog uri scheme
-    window.onload = function () {
+    <script>
+      // on load try to redirect to the app via liftlog uri scheme
+      window.onload = function () {
         try {
-            if (window.location.href.includes("app.liftlog.online"))
-                window.location.href = window.location.href
-                    .replace("https://", "liftlog://")
-                    .replace("http://", "liftlog://");
+          if (window.location.href.includes("app.liftlog.online"))
+            window.location.href = window.location.href
+              .replace("https://", "liftlog://")
+              .replace("http://", "liftlog://");
         } catch (e) {
-            // Not an app
+          // Not an app
         }
-    };
-</script>
-<!-- End Single Page Apps for GitHub Pages -->
+      };
+    </script>
+    <!-- End Single Page Apps for GitHub Pages -->
 
-<script src="/_framework/blazor.webassembly.js"></script>
-<script src="_content/LiftLog.Ui/app-utils.js"></script>
-<script src="_content/LiftLog.Ui/bundle.js"></script>
-<script src="/crypto-utils.js"></script>
-</body>
+    <script src="/_framework/blazor.webassembly.js"></script>
+    <script src="_content/LiftLog.Ui/app-utils.js"></script>
+    <script src="_content/LiftLog.Ui/bundle.js"></script>
+    <script src="/crypto-utils.js"></script>
+    <script src="/web-utils.js"></script>
+  </body>
 </html>
